recipe writing/recipe-writer/requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.21.0
requests>=2.25.0
